Move StatusDisplay to main project
