+++ LegacySignController/src/StatusDisplay/NullStatusDisplay.h
